+++ examples/playground/src/pages/registry.ts
@@ -40,6 +40,7 @@ import { kbdPage } from './kbd';
 import { statusPage } from './status';
 import { indicatorPage } from './indicator';
 import { statsPage } from './stats';
+import { timelinePage } from './timeline';
 import { progressPage } from './progress';
 import { sizeRampPage } from './size-ramp';
 import { extensibleRolesPage } from './extensible-roles';
@@ -109,6 +110,7 @@ export const pages: PageEntry[] = [
     statusPage,
     indicatorPage,
     statsPage,
+    timelinePage,
     progressPage,
     sizeRampPage,
     extensibleRolesPage,